Reuse item alternative logic for import node source
