+++ rm-portal/src/main/java/org/ow2/proactive_grid_cloud_portal/rm/client/nodesource/EditNodeSourceWindow.java
@@ -26,9 +26,7 @@
 package org.ow2.proactive_grid_cloud_portal.rm.client.nodesource;
 
 import java.util.LinkedHashMap;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Map;
 
 import org.ow2.proactive_grid_cloud_portal.rm.client.NodeSourceAction;
 import org.ow2.proactive_grid_cloud_portal.rm.client.NodeSourceConfiguration;
@@ -39,11 +37,8 @@ import com.smartgwt.client.widgets.Label;
 import com.smartgwt.client.widgets.form.DynamicForm;
 import com.smartgwt.client.widgets.form.fields.CheckboxItem;
 import com.smartgwt.client.widgets.form.fields.FormItem;
-import com.smartgwt.client.widgets.form.fields.RadioGroupItem;
 import com.smartgwt.client.widgets.form.fields.SpacerItem;
 import com.smartgwt.client.widgets.form.fields.TextAreaItem;
-import com.smartgwt.client.widgets.form.fields.TextItem;
-import com.smartgwt.client.widgets.form.fields.UploadItem;
 import com.smartgwt.client.widgets.layout.HLayout;
 
 
@@ -56,29 +51,18 @@ public class EditNodeSourceWindow extends NodeSourceWindow {
 
     public static final String WINDOW_TITLE = "Edit Node Source";
 
-    public static final String EDIT_OR_UPLOAD_FORM_ITEM_SUFFIX = ".modify";
-
-    public static final String EDIT_FORM_ITEM_SUFFIX = ".edit";
-
-    public static final String EDIT_RADIO_OPTION_NAME = "edit";
-
-    public static final String UPLOAD_RADIO_OPTION_NAME = "upload";
+    private NonTextualItemAternativeChoiceCreator nonTextualItemAternativeChoiceCreator;
 
     protected String nodeSourceName;
 
-    protected String focusedInfrastructurePluginName;
-
-    protected String focusedPolicyPluginName;
-
     public EditNodeSourceWindow(RMController controller, String nodeSourceName) {
-        super(controller, WINDOW_TITLE, "Retrieving current node source configuration");
-        this.nodeSourceName = nodeSourceName;
-        buildForm();
+        this(controller, WINDOW_TITLE, "Retrieving current node source configuration");
     }
 
     protected EditNodeSourceWindow(RMController controller, String nodeSourceName, String windowTitle) {
         super(controller, windowTitle, "Retrieving current node source configuration");
         this.nodeSourceName = nodeSourceName;
+        this.nonTextualItemAternativeChoiceCreator = new NonTextualItemAternativeChoiceCreator(this);
         buildForm();
     }
 
@@ -99,35 +83,18 @@ public class EditNodeSourceWindow extends NodeSourceWindow {
 
     @Override
     protected List<FormItem> handleNonTextualPluginField(PluginDescriptor plugin, PluginDescriptor.Field pluginField) {
-
-        FormItem chooseCredentialsFormItem;
-        List<FormItem> formItemsReplacingNonTextualFormItems = new LinkedList<>();
-
-        if (plugin.getPluginName().equals(this.focusedInfrastructurePluginName) ||
-            plugin.getPluginName().equals(this.focusedPolicyPluginName)) {
-
-            RadioGroupItem editOrUploadFormItem = createRadioItemToModifyPluginField(plugin, pluginField);
-            formItemsReplacingNonTextualFormItems.add(editOrUploadFormItem);
-
-            TextAreaItem previousValueItem = createTextItemPrefilledWithFileContent(plugin, pluginField);
-            formItemsReplacingNonTextualFormItems.add(previousValueItem);
-
-            chooseCredentialsFormItem = createUploadItemDisabled(plugin, pluginField);
-
-        } else {
-
-            chooseCredentialsFormItem = new UploadItem(plugin.getPluginName() + pluginField.getName(),
-                                                       pluginField.getName());
-        }
-
-        addCredentialsPickerIcon(pluginField, formItemsReplacingNonTextualFormItems, chooseCredentialsFormItem);
-
-        return formItemsReplacingNonTextualFormItems;
+        return nonTextualItemAternativeChoiceCreator.getModificationChoiceItemsForNonTextualFields(plugin, pluginField);
     }
 
     @Override
     protected void addButtonsToButtonsLayout(HLayout buttonsLayout) {
         buttonsLayout.setMembers(this.deployNowButton, this.saveAndKeepUndeployedButton, this.cancelButton);
+    }
+
+    @Override
+    protected void manageNodeSourceWindowItems(TextAreaItem nodeSourceNameItem, CheckboxItem nodesRecoverableItem) {
+        // we never allow the node source name to be modified
+        nodeSourceNameItem.disable();
     }
 
     protected void fetchNodeSourceConfigurationWithCallback(Label windowLabel, DynamicForm windowForm,
@@ -151,7 +118,7 @@ public class EditNodeSourceWindow extends NodeSourceWindow {
             PluginDescriptor focusedInfrastructurePlugin = nodeSourceConfiguration.getInfrastructurePluginDescriptor();
             this.focusedInfrastructurePluginName = focusedInfrastructurePlugin.getPluginName();
             this.allFormItems.add(this.infrastructureSelectItem);
-            addFocusedPluginValues(selectItemValues, focusedInfrastructurePlugin);
+            fillFocusedPluginValues(selectItemValues, focusedInfrastructurePlugin);
             handleAdditionalInfrastructureFormItems(selectItemValues, focusedInfrastructurePlugin);
             this.infrastructureSelectItem.setValueMap(selectItemValues);
             this.infrastructureSelectItem.setDefaultToFirstOption(true);
@@ -163,7 +130,7 @@ public class EditNodeSourceWindow extends NodeSourceWindow {
             PluginDescriptor focusedPolicyPlugin = nodeSourceConfiguration.getPolicyPluginDescriptor();
             this.focusedPolicyPluginName = focusedPolicyPlugin.getPluginName();
             this.allFormItems.add(this.policySelectItem);
-            addFocusedPluginValues(selectItemValues, focusedPolicyPlugin);
+            fillFocusedPluginValues(selectItemValues, focusedPolicyPlugin);
             handleAdditionalPolicyFormItems(selectItemValues, focusedPolicyPlugin);
             this.policySelectItem.setValueMap(selectItemValues);
             this.policySelectItem.setDefaultToFirstOption(true);
@@ -179,178 +146,6 @@ public class EditNodeSourceWindow extends NodeSourceWindow {
             windowForm.show();
 
         }, this.window::hide);
-    }
-
-    /**
-     * Allow sub classes to redefine a custom behavior regarding what to do,
-     * regarding policies, in addition to the focused policy plugin.
-     */
-    protected void handleAdditionalPolicyFormItems(Map<String, String> selectItemValues,
-            PluginDescriptor focusedPolicyPlugin) {
-        Map<String, PluginDescriptor> allSupportedPolicies = this.controller.getModel().getSupportedPolicies();
-        addPluginValuesToAllFormItemOtherThanFocused(selectItemValues, focusedPolicyPlugin, allSupportedPolicies);
-    }
-
-    /**
-     * Allow sub classes to redefine a custom behavior regarding what to do,
-     * regarding infrastructures, in addition to the focused infrastructure
-     * plugin.
-     */
-    protected void handleAdditionalInfrastructureFormItems(Map<String, String> selectItemValues,
-            PluginDescriptor focusedInfrastructurePlugin) {
-        Map<String, PluginDescriptor> allSupportedInfrastructures = this.controller.getModel()
-                                                                                   .getSupportedInfrastructures();
-        addPluginValuesToAllFormItemOtherThanFocused(selectItemValues,
-                                                     focusedInfrastructurePlugin,
-                                                     allSupportedInfrastructures);
-    }
-
-    /**
-     * Allow sub classes to rework the form items that have been added to the
-     * form.
-     */
-    protected List<FormItem> modifyFormItemsAfterCreation(PluginDescriptor focusedInfrastructurePlugin,
-            PluginDescriptor focusedPolicyPlugin) {
-
-        for (Map.Entry<String, List<FormItem>> entry : this.allFormItemsPerPlugin.entrySet()) {
-
-            hideNotFocusedFormItem(focusedInfrastructurePlugin, focusedPolicyPlugin, entry.getKey(), entry.getValue());
-        }
-
-        return this.allFormItems;
-    }
-
-    /**
-     * Allow sub classes to select the items they need to disable.
-     */
-    protected void manageNodeSourceWindowItems(TextAreaItem nodeSourceNameItem, CheckboxItem nodesRecoverableItem) {
-
-        // we never allow the node source name to be modified
-        nodeSourceNameItem.disable();
-    }
-
-    private FormItem createUploadItemDisabled(PluginDescriptor plugin, PluginDescriptor.Field pluginField) {
-
-        FormItem chooseCredentialsFormItem;
-        chooseCredentialsFormItem = new UploadItem(plugin.getPluginName() + pluginField.getName(), "");
-        chooseCredentialsFormItem.disable();
-        return chooseCredentialsFormItem;
-    }
-
-    private TextAreaItem createTextItemPrefilledWithFileContent(PluginDescriptor plugin,
-            PluginDescriptor.Field pluginField) {
-
-        TextAreaItem previousValueItem = new TextAreaItem(plugin.getPluginName() + pluginField.getName() +
-                                                          EDIT_FORM_ITEM_SUFFIX, "");
-
-        previousValueItem.setDefaultValue(pluginField.getValue());
-
-        return previousValueItem;
-    }
-
-    private RadioGroupItem createRadioItemToModifyPluginField(PluginDescriptor plugin,
-            PluginDescriptor.Field pluginField) {
-
-        String formItemPrefixName = plugin.getPluginName() + pluginField.getName();
-
-        LinkedHashMap<String, String> radioOptions = new LinkedHashMap<>();
-        radioOptions.put(EDIT_RADIO_OPTION_NAME, "In-line editing");
-        radioOptions.put(UPLOAD_RADIO_OPTION_NAME, "Upload new file");
-
-        RadioGroupItem editOrUploadFormItem = new RadioGroupItem(formItemPrefixName + EDIT_OR_UPLOAD_FORM_ITEM_SUFFIX,
-                                                                 pluginField.getName());
-        editOrUploadFormItem.setValueMap(radioOptions);
-        editOrUploadFormItem.setVertical(false);
-        editOrUploadFormItem.setDefaultValue(EDIT_RADIO_OPTION_NAME);
-
-        editOrUploadFormItem.addChangedHandler(changedEvent -> {
-
-            String radioValue = changedEvent.getValue().toString();
-            List<FormItem> formItemsForPlugin = this.allFormItemsPerPlugin.get(plugin.getPluginName());
-
-            if (radioValue.equals(EDIT_RADIO_OPTION_NAME)) {
-                enableEditInLine(formItemPrefixName, formItemsForPlugin);
-            } else if (radioValue.equals(UPLOAD_RADIO_OPTION_NAME)) {
-                enableUploadNewFile(formItemPrefixName, formItemsForPlugin);
-            }
-        });
-
-        return editOrUploadFormItem;
-    }
-
-    private void enableUploadNewFile(String pluginFieldName, List<FormItem> formItemsForPlugin) {
-
-        formItemsForPlugin.stream()
-                          .filter(formItem -> formItem.getName().startsWith(pluginFieldName))
-                          .forEach(formItem -> enableUploadNewFile(pluginFieldName, formItem));
-    }
-
-    private void enableUploadNewFile(String pluginFieldName, FormItem formItem) {
-
-        if (formItem.getName().equals(pluginFieldName)) {
-            formItem.enable();
-
-        } else if (formItem.getName().startsWith(pluginFieldName) &&
-                   formItem.getName().endsWith(EDIT_FORM_ITEM_SUFFIX)) {
-            formItem.disable();
-        }
-    }
-
-    private void enableEditInLine(String pluginFieldName, List<FormItem> formItemsForPlugin) {
-
-        formItemsForPlugin.stream()
-                          .filter(formItem -> formItem.getName().startsWith(pluginFieldName))
-                          .forEach(formItem -> enableEditInLine(pluginFieldName, formItem));
-    }
-
-    private void enableEditInLine(String pluginFieldName, FormItem formItem) {
-
-        if (formItem.getName().equals(pluginFieldName)) {
-            formItem.disable();
-
-        } else if (formItem.getName().startsWith(pluginFieldName) &&
-                   formItem.getName().endsWith(EDIT_FORM_ITEM_SUFFIX)) {
-            formItem.enable();
-        }
-    }
-
-    private void hideNotFocusedFormItem(PluginDescriptor focusedInfrastructurePlugin,
-            PluginDescriptor focusedPolicyPlugin, String pluginName, List<FormItem> formItemsForPlugin) {
-
-        if (!pluginName.equals(focusedInfrastructurePlugin.getPluginName()) &&
-            !pluginName.equals(focusedPolicyPlugin.getPluginName())) {
-
-            for (FormItem formItem : formItemsForPlugin) {
-                formItem.hide();
-            }
-        }
-    }
-
-    private void addFocusedPluginValues(LinkedHashMap<String, String> selectItemValues,
-            PluginDescriptor focusedPlugin) {
-        String pluginShortName = getPluginShortName(focusedPlugin);
-        selectItemValues.put(focusedPlugin.getPluginName(), pluginShortName);
-
-        List<FormItem> pluginFormItems = getPrefilledFormItems(focusedPlugin);
-        this.allFormItems.addAll(pluginFormItems);
-        this.allFormItemsPerPlugin.put(focusedPlugin.getPluginName(), pluginFormItems);
-    }
-
-    private void addPluginValuesToAllFormItemOtherThanFocused(Map<String, String> selectItemValues,
-            PluginDescriptor focusedPlugin, Map<String, PluginDescriptor> allPluginDescriptors) {
-
-        for (Map.Entry<String, PluginDescriptor> entry : allPluginDescriptors.entrySet()) {
-
-            PluginDescriptor plugin = entry.getValue();
-
-            if (!plugin.getPluginName().equals(focusedPlugin.getPluginName())) {
-
-                selectItemValues.put(plugin.getPluginName(), getPluginShortName(plugin));
-                List<FormItem> currentPluginFormItems = getPrefilledFormItems(plugin);
-                this.allFormItems.addAll(currentPluginFormItems);
-                this.allFormItemsPerPlugin.put(plugin.getPluginName(), currentPluginFormItems);
-            }
-        }
     }
 
 }
